+++ apps/web/src/app/[locale]/(space)/account/billing/page.tsx
@@ -29,7 +29,7 @@ import {
 } from "@/components/empty-state";
 import { PayWallDialog } from "@/components/pay-wall-dialog";
 import { Trans } from "@/components/trans";
-import { loadPaymentMethods, loadSubscription } from "@/data/space";
+import { loadPaymentMethods, loadSubscription } from "@/features/spaces/data";
 import { FormattedDateTime } from "@/features/timezone/client/formatted-date-time";
 import { getTranslation } from "@/i18n/server";
 import { isSelfHosted } from "@/utils/constants";
