+++ apps/web/src/app/[locale]/(space)/components/sidebar/space-sidebar.tsx
@@ -15,11 +15,11 @@ import { SpaceSidebarMenu } from "@/app/[locale]/(space)/components/sidebar/spac
 import { LogoLink } from "@/app/components/logo-link";
 import { loadCurrentUserSpace } from "@/auth/data";
 import { Trans } from "@/components/trans";
-import { loadSpaces } from "@/data/space";
 import { FeedbackToggle } from "@/features/feedback/components/feedback-toggle";
 import { SpaceDropdown } from "@/features/spaces/components/space-dropdown";
 import { SpaceIcon } from "@/features/spaces/components/space-icon";
 import { isSpacesEnabled } from "@/features/spaces/constants";
+import { loadSpaces } from "@/features/spaces/data";
 import { UpgradeButton } from "../upgrade-button";
 import { NavUser } from "./nav-user";
