+++ apps/web/src/auth/data.ts
@@ -3,7 +3,7 @@ import "server-only";
 import { prisma } from "@rallly/database";
 import { notFound, redirect } from "next/navigation";
 import { cache } from "react";
-import { createSpaceDTO } from "@/data/space";
+import { createSpaceDTO } from "@/features/spaces/data";
 import { createUserDTO } from "@/features/user/data";
 import { AppError } from "@/lib/errors";
 import { auth } from "@/next-auth";
